circular import fix, general.py added
